Factor out ui_panel.rs
